TravelCompanion/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain-google-genai>=2.1.3",
    "python-dotenv>=1.1.0",
    "streamlit-folium>=0.25.0",
    "streamlit>=1.44.1",
    "folium>=0.19.5",
    "geopy>=2.4.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "openai>=1.75.0",
]
